+++ docs/ReleaseNotes.md
@@ -14,6 +14,7 @@ The Guava dependency version has been updated to 31.1. Projects may need to chec
 <!--
 // begin next release
 ### NEXT_RELEASE
+* **Feature** Supports bitset_contains in Lucene query
 
 * **Bug fix** Fix 1 [(Issue #NNN)](https://github.com/FoundationDB/fdb-record-layer/issues/NNN)
 * **Bug fix** Fix 2 [(Issue #NNN)](https://github.com/FoundationDB/fdb-record-layer/issues/NNN)